refactor(activities): split api.go god object per active-package template (#575 B5)
Pure file moves: api.go (2005 L) → wiring-only (79 L) + activity_handlers
/ schedule_handlers / supervisor_handlers / enrollment_handlers /
helpers + types.go. Route table unchanged (B0 golden). Complexity
ratchet entry re-pathed.
